fixed to produce outcomes for floats and integers

+++ node-require/calculate.js
@@ -4,17 +4,17 @@ const divide = require('./divide');
 const multiply = require('./multiply');
 
 if (process.argv[3] === 'plus') {
-  console.log(add(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log(add(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 }
 
 if (process.argv[3] === 'minus') {
-  console.log(subtract(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log(subtract(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 }
 
 if (process.argv[3] === 'over') {
-  console.log(divide(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log(divide(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 }
 
 if (process.argv[3] === 'times') {
-  console.log(multiply(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log(multiply(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 }
